fix pr_build

+++ scripts/build.py
@@ -97,7 +97,8 @@ if os.path.exists(r'target'):
             shutil.copy(r'readme.txt', r'releases/readme.txt')
             shutil.copy(r'Ultimate S Setup Tool.exe', r'releases/Ultimate S Setup Tool.exe')
             shutil.copy(r'Ultimate S Setup Tool.py', r'releases/Ultimate S Setup Tool.py')
-            shutil.copytree(r'resources', r'releases/resources')
+            if os.path.exists(r'resources'):
+                shutil.copytree(r'resources', r'releases/resources')
 
             if os.path.exists(r'releases/Ultimate S Arcropolis.zip'):
                 os.remove(r'releases/Ultimate S Arcropolis.zip')
